Update README.md
At the most basic level, a chatbot is a computer program that simulates and processes human conversation (either written or spoken), allowing humans to interact with digital devices as if they were communicating with a real person. User communicate with chatbot via the chat interface or by voice, like how they talk to real person. Chatbots interpret and process user's words or phrases and give an instant pre-set answer.
. Food ordering chat bot 
. Ordering through text 
. Can be converted to voice 
. Configuration text is through files